requirements.txt
# This txt records the packs needed
# conda install pytorch==1.8.0 torchvision torchaudio cudatoolkit=10.2
# conda install pytorch==1.8.0 torchvision==0.9.0 torchaudio==0.8.0 cpuonly
pyqt5
matplotlib~=3.7.1
scikit-learn
timm
albumentations
pillow==8.4.0
opencv-python~=4.7.0.72
pandas
tqdm~=4.65.0
torchstat
requests~=2.31.0
numpy~=1.24.3
torchvision~=0.11.0
imageio~=2.31.3
